# api/v1/cluster_manager_server.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/v1/cluster_manager_server.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "ClustersService"
    },
    {
      "name": "ClustersInternalService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/clusters": {
      "get": {
        "operationId": "ClustersService_ListClusters",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListClustersResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "ClustersService"
        ]
      },
      "post": {
        "operationId": "ClustersService_CreateCluster",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Cluster"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateClusterRequest"
            }
          }
        ],
        "tags": [
          "ClustersService"
        ]
      }
    },
    "/v1/clusters/{id}": {
      "get": {
        "operationId": "ClustersService_GetCluster",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Cluster"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "ClustersService"
        ]
      },
      "delete": {
        "operationId": "ClustersService_DeleteCluster",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteClusterResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "ClustersService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1Cluster": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "registrationKey": {
          "type": "string",
          "description": "registration_key is the key used to register the cluster. Only populated\nwhen a new cluster is created or from internal gRPC service."
        },
        "object": {
          "type": "string"
        },
        "componentStatuses": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/v1ComponentStatus"
          },
          "description": "components_statuses is keyed by component name."
        }
      }
    },
    "v1ComponentStatus": {
      "type": "object",
      "properties": {
        "isHealthy": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "v1CreateClusterRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
    "v1DeleteClusterResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "object": {
          "type": "string"
        },
        "deleted": {
          "type": "boolean"
        }
      }
    },
    "v1InternalCluster": {
      "type": "object",
      "properties": {
        "cluster": {
          "$ref": "#/definitions/v1Cluster"
        },
        "tenantId": {
          "type": "string"
        }
      }
    },
    "v1ListClustersResponse": {
      "type": "object",
      "properties": {
        "object": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1Cluster"
          }
        }
      }
    },
    "v1ListInternalClustersResponse": {
      "type": "object",
      "properties": {
        "clusters": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1InternalCluster"
          }
        }
      }
    }
  }
}